Move edittagdialog.py to metadatabox/ and use relative imports
