+++ picard/ui/metadatabox/edittagdialog.py
